requirements.txt
# ============================================
# Chest X-Ray Pneumonia Detection
# ResNet50 + Stacking Ensemble (SVM + RF + XGBoost)
# Python 3.10+
# ============================================

# === Deep Learning & Feature Extraction ===
tensorflow>=2.15.0
keras>=3.0.0

# === Machine Learning ===
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# === Image Processing ===
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# === Data & Computation ===
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
tqdm>=4.65.0

# === Visualization ===
matplotlib>=3.7.0
plotly>=5.20.0
dash>=2.17.0

# === Web App (FastAPI) ===
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Jinja2>=3.1.2
